+++ .pnp.cjs
@@ -6998,7 +6998,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@chainlink/ea-bootstrap", "workspace:packages/core/bootstrap"],
             ["@chainlink/ea-test-helpers", "workspace:packages/core/test-helpers"],
             ["@chainlink/types", "workspace:packages/core/types/@chainlink"],
-            ["@synthetixio/contracts-interface", "npm:2.61.0"],
             ["@types/jest", "npm:27.0.3"],
             ["@types/node", "npm:16.11.19"],
             ["@types/supertest", "npm:2.0.11"],
@@ -10982,19 +10981,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "linkType": "HARD",
         }]
       ]],
-      ["@synthetixio/contracts-interface", [
-        ["npm:2.61.0", {
-          "packageLocation": "./.yarn/cache/@synthetixio-contracts-interface-npm-2.61.0-1c06ad9b46-575f52f727.zip/node_modules/@synthetixio/contracts-interface/",
-          "packageDependencies": [
-            ["@synthetixio/contracts-interface", "npm:2.61.0"],
-            ["ethers", "npm:5.5.3"],
-            ["lodash", "npm:4.17.19"],
-            ["synthetix", "npm:2.61.0"],
-            ["web3-utils", "npm:1.2.11"]
-          ],
-          "linkType": "HARD",
-        }]
-      ]],
       ["@szmarczak/http-timer", [
         ["npm:1.1.2", {
           "packageLocation": "./.yarn/cache/@szmarczak-http-timer-npm-1.1.2-ea82ca2d55-4d9158061c.zip/node_modules/@szmarczak/http-timer/",
@@ -12974,13 +12960,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         }]
       ]],
       ["ansi-escapes", [
-        ["npm:3.2.0", {
-          "packageLocation": "./.yarn/cache/ansi-escapes-npm-3.2.0-a9d573100e-0f94695b67.zip/node_modules/ansi-escapes/",
-          "packageDependencies": [
-            ["ansi-escapes", "npm:3.2.0"]
-          ],
-          "linkType": "HARD",
-        }],
         ["npm:4.3.2", {
           "packageLocation": "./.yarn/cache/ansi-escapes-npm-4.3.2-3ad173702f-93111c4218.zip/node_modules/ansi-escapes/",
           "packageDependencies": [
@@ -15263,14 +15242,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         }]
       ]],
       ["cli-cursor", [
-        ["npm:2.1.0", {
-          "packageLocation": "./.yarn/cache/cli-cursor-npm-2.1.0-3920629c9c-d88e97bfda.zip/node_modules/cli-cursor/",
-          "packageDependencies": [
-            ["cli-cursor", "npm:2.1.0"],
-            ["restore-cursor", "npm:2.0.0"]
-          ],
-          "linkType": "HARD",
-        }],
         ["npm:3.1.0", {
           "packageLocation": "./.yarn/cache/cli-cursor-npm-3.1.0-fee1e46b5e-2692784c6c.zip/node_modules/cli-cursor/",
           "packageDependencies": [
@@ -15287,15 +15258,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["cli-truncate", "npm:2.1.0"],
             ["slice-ansi", "npm:3.0.0"],
             ["string-width", "npm:4.2.2"]
-          ],
-          "linkType": "HARD",
-        }]
-      ]],
-      ["cli-width", [
-        ["npm:2.2.1", {
-          "packageLocation": "./.yarn/cache/cli-width-npm-2.2.1-4bdb77393c-3c21b897a2.zip/node_modules/cli-width/",
-          "packageDependencies": [
-            ["cli-width", "npm:2.2.1"]
           ],
           "linkType": "HARD",
         }]
@@ -18791,16 +18753,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "linkType": "HARD",
         }]
       ]],
-      ["figures", [
-        ["npm:2.0.0", {
-          "packageLocation": "./.yarn/cache/figures-npm-2.0.0-f2db814eec-081beb16ea.zip/node_modules/figures/",
-          "packageDependencies": [
-            ["figures", "npm:2.0.0"],
-            ["escape-string-regexp", "npm:1.0.5"]
-          ],
-          "linkType": "HARD",
-        }]
-      ]],
       ["file-entry-cache", [
         ["npm:6.0.1", {
           "packageLocation": "./.yarn/cache/file-entry-cache-npm-6.0.1-31965cf0af-f49701feaa.zip/node_modules/file-entry-cache/",
@@ -20639,54 +20591,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "packageLocation": "./.yarn/cache/ini-npm-2.0.0-28f7426761-e7aadc5fb2.zip/node_modules/ini/",
           "packageDependencies": [
             ["ini", "npm:2.0.0"]
-          ],
-          "linkType": "HARD",
-        }]
-      ]],
-      ["inquirer", [
-        ["npm:6.5.2", {
-          "packageLocation": "./.yarn/cache/inquirer-npm-6.5.2-4f6408c247-175ad4cd1e.zip/node_modules/inquirer/",
-          "packageDependencies": [
-            ["inquirer", "npm:6.5.2"],
-            ["ansi-escapes", "npm:3.2.0"],
-            ["chalk", "npm:2.4.2"],
-            ["cli-cursor", "npm:2.1.0"],
-            ["cli-width", "npm:2.2.1"],
-            ["external-editor", "npm:3.1.0"],
-            ["figures", "npm:2.0.0"],
-            ["lodash", "npm:4.17.21"],
-            ["mute-stream", "npm:0.0.7"],
-            ["run-async", "npm:2.4.1"],
-            ["rxjs", "npm:6.6.7"],
-            ["string-width", "npm:2.1.1"],
-            ["strip-ansi", "npm:5.2.0"],
-            ["through", "npm:2.3.8"]
-          ],
-          "linkType": "HARD",
-        }]
-      ]],
-      ["inquirer-list-search-prompt", [
-        ["npm:1.0.2", {
-          "packageLocation": "./.yarn/cache/inquirer-list-search-prompt-npm-1.0.2-1b893c38a0-0bf538a27e.zip/node_modules/inquirer-list-search-prompt/",
-          "packageDependencies": [
-            ["inquirer-list-search-prompt", "npm:1.0.2"]
-          ],
-          "linkType": "SOFT",
-        }],
-        ["virtual:ddef24006e74e32e71d4488e17885e419436c5978dd4d5cb0424918d4007c689e32bddd7f2692a329aa538eaea195cee3c847349b12663bdab910caaabd4c417#npm:1.0.2", {
-          "packageLocation": "./.yarn/__virtual__/inquirer-list-search-prompt-virtual-e2fb598d2f/0/cache/inquirer-list-search-prompt-npm-1.0.2-1b893c38a0-0bf538a27e.zip/node_modules/inquirer-list-search-prompt/",
-          "packageDependencies": [
-            ["inquirer-list-search-prompt", "virtual:ddef24006e74e32e71d4488e17885e419436c5978dd4d5cb0424918d4007c689e32bddd7f2692a329aa538eaea195cee3c847349b12663bdab910caaabd4c417#npm:1.0.2"],
-            ["@types/inquirer", null],
-            ["ansi-escapes", "npm:3.2.0"],
-            ["chalk", "npm:2.4.2"],
-            ["figures", "npm:2.0.0"],
-            ["inquirer", "npm:6.5.2"],
-            ["run-async", "npm:2.4.1"]
-          ],
-          "packagePeers": [
-            "@types/inquirer",
-            "inquirer"
           ],
           "linkType": "HARD",
         }]
@@ -23401,15 +23305,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "linkType": "HARD",
         }]
       ]],
-      ["js-levenshtein", [
-        ["npm:1.1.6", {
-          "packageLocation": "./.yarn/cache/js-levenshtein-npm-1.1.6-ab883e61a3-409f052a7f.zip/node_modules/js-levenshtein/",
-          "packageDependencies": [
-            ["js-levenshtein", "npm:1.1.6"]
-          ],
-          "linkType": "HARD",
-        }]
-      ]],
       ["js-sha3", [
         ["npm:0.5.7", {
           "packageLocation": "./.yarn/cache/js-sha3-npm-0.5.7-e4e4cb1ee3-973a28ea4b.zip/node_modules/js-sha3/",
@@ -24796,13 +24691,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         }]
       ]],
       ["lodash", [
-        ["npm:4.17.19", {
-          "packageLocation": "./.yarn/cache/lodash-npm-4.17.19-fa87a8e838-a5cc77099f.zip/node_modules/lodash/",
-          "packageDependencies": [
-            ["lodash", "npm:4.17.19"]
-          ],
-          "linkType": "HARD",
-        }],
         ["npm:4.17.21", {
           "packageLocation": "./.yarn/cache/lodash-npm-4.17.21-6382451519-eb835a2e51.zip/node_modules/lodash/",
           "packageDependencies": [
@@ -26368,15 +26256,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "linkType": "HARD",
         }]
       ]],
-      ["mute-stream", [
-        ["npm:0.0.7", {
-          "packageLocation": "./.yarn/cache/mute-stream-npm-0.0.7-22b59a65dd-a9d4772c1c.zip/node_modules/mute-stream/",
-          "packageDependencies": [
-            ["mute-stream", "npm:0.0.7"]
-          ],
-          "linkType": "HARD",
-        }]
-      ]],
       ["nan", [
         ["npm:2.14.2", {
           "packageLocation": "./.yarn/unplugged/nan-npm-2.14.2-e3ede8ce5d/node_modules/nan/",
@@ -27207,14 +27086,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         }]
       ]],
       ["onetime", [
-        ["npm:2.0.1", {
-          "packageLocation": "./.yarn/cache/onetime-npm-2.0.1-6c39ecc911-bb44015ac7.zip/node_modules/onetime/",
-          "packageDependencies": [
-            ["onetime", "npm:2.0.1"],
-            ["mimic-fn", "npm:1.2.0"]
-          ],
-          "linkType": "HARD",
-        }],
         ["npm:5.1.2", {
           "packageLocation": "./.yarn/cache/onetime-npm-5.1.2-3ed148fa42-2478859ef8.zip/node_modules/onetime/",
           "packageDependencies": [
@@ -29536,15 +29407,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         }]
       ]],
       ["restore-cursor", [
-        ["npm:2.0.0", {
-          "packageLocation": "./.yarn/cache/restore-cursor-npm-2.0.0-80278eb6b7-482e13d02d.zip/node_modules/restore-cursor/",
-          "packageDependencies": [
-            ["restore-cursor", "npm:2.0.0"],
-            ["onetime", "npm:2.0.1"],
-            ["signal-exit", "npm:3.0.3"]
-          ],
-          "linkType": "HARD",
-        }],
         ["npm:3.1.0", {
           "packageLocation": "./.yarn/cache/restore-cursor-npm-3.1.0-52c5a4c98f-f877dd8741.zip/node_modules/restore-cursor/",
           "packageDependencies": [
@@ -29722,15 +29584,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["rsv-signature", "npm:1.1.0"],
             ["bn.js", "npm:4.12.0"],
             ["enc-utils", "npm:2.2.3"]
-          ],
-          "linkType": "HARD",
-        }]
-      ]],
-      ["run-async", [
-        ["npm:2.4.1", {
-          "packageLocation": "./.yarn/cache/run-async-npm-2.4.1-a94bb90861-a2c88aa15d.zip/node_modules/run-async/",
-          "packageDependencies": [
-            ["run-async", "npm:2.4.1"]
           ],
           "linkType": "HARD",
         }]
@@ -31390,21 +31243,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["synthetix", "npm:2.45.3"],
             ["abi-decoder", "npm:2.3.0"],
             ["commander", "npm:5.1.0"],
-            ["pretty-error", "npm:2.1.2"],
-            ["solidity-parser-antlr", "npm:0.4.11"],
-            ["web3-utils", "npm:1.2.2"]
-          ],
-          "linkType": "HARD",
-        }],
-        ["npm:2.61.0", {
-          "packageLocation": "./.yarn/cache/synthetix-npm-2.61.0-ddef24006e-d61130c947.zip/node_modules/synthetix/",
-          "packageDependencies": [
-            ["synthetix", "npm:2.61.0"],
-            ["abi-decoder", "npm:2.3.0"],
-            ["commander", "npm:8.1.0"],
-            ["inquirer", "npm:6.5.2"],
-            ["inquirer-list-search-prompt", "virtual:ddef24006e74e32e71d4488e17885e419436c5978dd4d5cb0424918d4007c689e32bddd7f2692a329aa538eaea195cee3c847349b12663bdab910caaabd4c417#npm:1.0.2"],
-            ["js-levenshtein", "npm:1.1.6"],
             ["pretty-error", "npm:2.1.2"],
             ["solidity-parser-antlr", "npm:0.4.11"],
             ["web3-utils", "npm:1.2.2"]
@@ -33601,21 +33439,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         }]
       ]],
       ["web3-utils", [
-        ["npm:1.2.11", {
-          "packageLocation": "./.yarn/cache/web3-utils-npm-1.2.11-618f0ef456-1e43235963.zip/node_modules/web3-utils/",
-          "packageDependencies": [
-            ["web3-utils", "npm:1.2.11"],
-            ["bn.js", "npm:4.12.0"],
-            ["eth-lib", "npm:0.2.8"],
-            ["ethereum-bloom-filters", "npm:1.0.10"],
-            ["ethjs-unit", "npm:0.1.6"],
-            ["number-to-bn", "npm:1.7.0"],
-            ["randombytes", "npm:2.1.0"],
-            ["underscore", "npm:1.9.1"],
-            ["utf8", "npm:3.0.0"]
-          ],
-          "linkType": "HARD",
-        }],
         ["npm:1.2.2", {
           "packageLocation": "./.yarn/cache/web3-utils-npm-1.2.2-30d9c9754b-bbd1189b7b.zip/node_modules/web3-utils/",
           "packageDependencies": [