+++ packages/sources/synthetix-debt-pool/src/endpoint/debt.ts
@@ -1,12 +1,10 @@
 import { AdapterError } from '@chainlink/ea-bootstrap'
 import { ExecuteWithConfig } from '@chainlink/types'
-import { BigNumber } from 'ethers'
-import {
-  getChainSynthetixInstance,
-  getDataFromAcrossChains,
-  inputParameters as commonInputParameters,
-} from '../commons'
+import { BigNumber, ethers } from 'ethers'
+import { getDataFromAcrossChains, inputParameters as commonInputParameters } from '../utils'
 import { Config } from '../config'
+import { getContractAddress } from '../utils'
+import { DEBT_CACHE_ABI } from './abi'
 
 // Needs to be exported so that doc generator script works
 export const inputParameters = commonInputParameters
@@ -22,9 +20,21 @@ const getTotalDebtIssued = async (
 ): Promise<BigNumber> => {
   const chainResponses = await Promise.all(
     chainsToQuery.map(async (network): Promise<BigNumber> => {
-      const snxjs = getChainSynthetixInstance(network, jobRunID, config)
+      if (!config.chains[network])
+        throw new AdapterError({
+          jobRunID,
+          statusCode: 500,
+          message: `Chain ${network} not configured`,
+        })
+      const networkProvider = new ethers.providers.JsonRpcProvider(config.chains[network].rpcURL)
       try {
-        const [debtIssued] = await snxjs.contracts.DebtCache.currentDebt()
+        const debtCacheAddress = await getContractAddress(
+          networkProvider,
+          config.chains[network].chainAddressResolverAddress,
+          'DebtCache',
+        )
+        const debtCache = new ethers.Contract(debtCacheAddress, DEBT_CACHE_ABI, networkProvider)
+        const [debtIssued] = await debtCache.currentDebt()
         return debtIssued
       } catch (e) {
         throw new AdapterError({