+++ packages/composites/outlier-detection/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/outlier-detection-adapter
 
+## 1.0.61
+
+### Patch Changes
+
+- @chainlink/ea@1.0.61
+
+## 1.0.60
+
+### Patch Changes
+
+- @chainlink/ea@1.0.60
+
 ## 1.0.59
 
 ### Patch Changes